Changes to hitobject.js;Converted to Module.
