+++ .gitignore
@@ -1,11 +1,2 @@
-*.log
-*.swp
-/out/*
-tags
-/test_suite
-/test_suite.json
-/package-lock.json
-/docs
 /node_modules
-*.tar.xz
-*.tar.xz.*
+Tests.js
