+++ src/hitobjects/index.js
@@ -1,0 +1,11 @@
+const {circle} = require('./circle');
+const {slider} = require('./slider');
+const {hitobject} = require('./hitobject');
+const {objtypes} = require('./objtypes')
+
+module.exports = {
+  objtypes: objtypes,
+  circle: circle,
+  slider: slider,
+  hitobject: hitobject,
+};
